ss: reduce logging and state type

+++ lib/secure-streams/system/captive-portal-detect/captive-portal-detect.c
@@ -41,7 +41,7 @@ ss_cpd_state(void *userobj, void *sh, lws_ss_constate_t state,
 	ss_cpd_t *m = (ss_cpd_t *)userobj;
 	struct lws_context *cx = (struct lws_context *)m->opaque_data;
 
-	lwsl_ss_info(m->ss, "%s, ord 0x%x\n", lws_ss_state_name((int)state),
+	lwsl_ss_info(m->ss, "%s, ord 0x%x\n", lws_ss_state_name(state),
 		  (unsigned int)ack);
 
 	switch (state) {
